Execute creat-react-native-module again to set up pod correctly.
