+++ example/ios/Podfile
@@ -27,8 +27,6 @@ target 'example' do
   pod 'glog', :podspec => '../node_modules/react-native/third-party-podspecs/glog.podspec'
   pod 'Folly', :podspec => '../node_modules/react-native/third-party-podspecs/Folly.podspec'
 
-	# pod 'react-native-use-health-kit', :path => '../../react-native-use-health-kit.podspec'
-
   target 'exampleTests' do
     inherit! :search_paths
     # Pods for testing
