--- example/ios/UseHealthKit/UseHealthKit-Bridging-Header.h
@@ -1,1 +1,0 @@
-#import "React/RCTBridgeModule.h"
