--- example/ios/UseHealthKit/UseHealthKit.h
@@ -1,5 +1,0 @@
-#import <React/RCTViewManager.h>
-
-@interface UseHealthKit : RCTViewManager
-
-@end
