+++ example/ios/dummy-for-bridge.swift
@@ -1,0 +1,9 @@
+//
+//  dummy-for-bridge.swift
+//  example
+//
+//  Created by Yoshifumi Kanno on 2019/09/26.
+//  Copyright © 2019 Facebook. All rights reserved.
+//
+
+import Foundation
