+++ ios/UseHealthKit.m
